+++ wifit_core/__init__.py
@@ -3,7 +3,7 @@
 A modular, testable implementation of WPS security assessment tools.
 """
 
-__version__ = "3.0.0-rc.1"
+__version__ = "3.0.0-rc.2"
 
 from .models import (
     AccessPoint,
@@ -14,31 +14,49 @@ from .models import (
     WPSVersion,
     WPSVersionEvidence,
 )
+from .pin_generator import PINGenerator, MACAddress, generate_pin, get_likely_pins, wps_checksum
+from .pixie_dust import PixiewpsParameters, PixiewpsResult, run_pixiewps
 from .platform import PlatformManager, WirelessInterface
 from .process_manager import ProcessManager, ProcessSnapshot
 from .reporter import ReportFormat, ResultReporter
 from .runner import CommandRunner, CommandResult
 from .scanner import WiFiScanner, parse_iw_scan
 from .vulnerability import WSCVulnerabilityDatabase, annotate_access_points
+from .wps_attack import WPASupplicantController, try_pin_attack, try_pbc_attack
+from .wps_bruteforce import BruteforceSession, BruteforceProgress, enumerate_all_pins
 
 __all__ = [
     "AccessPoint",
     "AttackMethod",
     "AttackOutcome",
     "AttackResult",
+    "BruteforceProgress",
+    "BruteforceSession",
     "CommandResult",
     "CommandRunner",
+    "MACAddress",
+    "PINGenerator",
+    "PixiewpsParameters",
+    "PixiewpsResult",
     "PlatformManager",
     "ProcessManager",
     "ProcessSnapshot",
     "ReportFormat",
     "ResultReporter",
     "SecurityMode",
+    "WPASupplicantController",
     "WiFiScanner",
     "WPSVersion",
     "WPSVersionEvidence",
     "WSCVulnerabilityDatabase",
     "WirelessInterface",
     "annotate_access_points",
+    "enumerate_all_pins",
+    "generate_pin",
+    "get_likely_pins",
     "parse_iw_scan",
+    "run_pixiewps",
+    "try_pbc_attack",
+    "try_pin_attack",
+    "wps_checksum",
 ]